Refactor/tidy up - remove redundant file write, move llama-server logic to separate file.
